Implement spin-table release method for secondary cores.
- Enable SMP as well as DDB and KDB
- Calculate CPU ID manually to support platforms without PSCI where there is
no possibility to pass the ID to mpentry(). CPU affinity is used to calculate
this ID, as well as information from FDT regarding number of CPUs in a cluster
on a given platform, etc.
- Include opt_platform.h to detect FDT option with #ifdef
- Remove psci.h include in mp_machdep.c (TODO: recognize PSCI option)
- Add option 'DO_SEV_LATER' to do 'sev' only after all CPU release addresses
have been initialized. This should not be necessary as long as reserving regions
works correctly, but is currently enabled as a precaution.

Currently release addresses for APs are hardcoded and specific to APM XGene.
As an improvement they should be parsed from FDT and stored. The release method
should also be chosen based on FDT.