Remove unnecessary incrementation of IPI number before sending
Number of IPI to be sent was incremented by 16 in ipi_cpu() and ipi_selected().
GICv2 and GICv3 registers for sending SGI expect interrupt number on bits 3:0,
therefore the addition was unnecessary. Bit 4 of those registers is RESERVED
so nothing important was overwritten.
This change should not affect any functionality.