Fix the write-sequence
In some I2C controllers, to write data to slave device, it is necessary
to send an internal register address and data to be written in a single
data write sequence.

This is because the first data byte of a write operation is always written
to the address pointer register, so it always contains an internal address
register. The second and next bytes are the data that are written to the
register selected by the address pointer register.

During write operation, the STOP condition cannot occur between consecutive
bytes. Controller generates the START or STOP condition automatically depending
on TX or RX buffer status and it is not possible to set it manually.

This modification changes a write sequence. Then Master device sends the slave
address and waits fo acknowledge. If this happens then the master sends the
internal register address and data to send it in data write sequence. Next
the controller waits fo acknowledge and automatically generates a STOP condition
on the bus.